requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
pandas>=1.3.3
yfinance>=0.1.63
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.1 